More clang fixes.  Thanks Arne #rb arne.schober
[CL 4737641 by Marcus Wassmer in Dev-Rendering branch]

+++ Engine/Source/Runtime/D3D12RHI/Private/D3D12Allocation.h
@@ -403,7 +403,7 @@ private:
 
 		Count,
 	};
-	FD3D12DefaultBufferPool* DefaultBufferPools[EBufferPool::Count];
+	FD3D12DefaultBufferPool* DefaultBufferPools[(uint32)EBufferPool::Count];
 
 	inline EBufferPool GetBufferPool(D3D12_RESOURCE_FLAGS Flags) const
 	{
